changed read ids to i64

+++ src/codon_structs.rs
@@ -68,7 +68,7 @@ pub trait Translations {
     fn find_mutations(&self,
                       gene: &bio::io::gff::Record,
                       variant_abundances: &HashMap<i32, BTreeMap<String, f64>>,
-                      indels: &HashMap<i32, BTreeMap<String, BTreeSet<i32>>>,
+                      indels: &HashMap<i32, BTreeMap<String, BTreeSet<i64>>>,
                       ref_sequence: &Vec<u8>,
                       depth: &Vec<f64>) -> f32;
 }
@@ -118,7 +118,7 @@ impl Translations for CodonTable {
     fn find_mutations(&self,
                       gene: &bio::io::gff::Record,
                       variant_abundances: &HashMap<i32, BTreeMap<String, f64>>,
-                      indels: &HashMap<i32, BTreeMap<String, BTreeSet<i32>>>,
+                      indels: &HashMap<i32, BTreeMap<String, BTreeSet<i64>>>,
                       ref_sequence: &Vec<u8>,
                       _depth: &Vec<f64>) -> f32 {
         let strand = gene.strand().expect("No strandedness found");
